requirements.txt
Django==5.2.5
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
pymongo==4.6.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
bcrypt==4.1.2
bakong-khqr>=1.0.0
qrcode[pil]>=7.4.2
